fix housing object init structure, game no longer crashes/doesn't load

+++ src/world/Inventory/HousingItem.cpp
@@ -8,12 +8,12 @@ Sapphire::Inventory::HousingItem::HousingItem( uint64_t uId, uint32_t catalogId,
   m_reservedFlag = 1092616192; // wat?
 }
 
-uint16_t Sapphire::Inventory::HousingItem::getRot() const
+float Sapphire::Inventory::HousingItem::getRot() const
 {
   return m_rotation;
 }
 
-void Sapphire::Inventory::HousingItem::setRot( uint16_t rot )
+void Sapphire::Inventory::HousingItem::setRot( float rot )
 {
   m_rotation = rot;
 }
